third_party/WebKit/LayoutTests/css3/flexbox/overflow-auto-resizes-correctly.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
.vflex {
  display: flex;
  flex-direction: column;
  max-width: 200px;
  margin: 10px 0 10px 0;
}
.hflex {
  display: flex;
  flex-direction: row;
  max-height: 200px;
  margin: 10px 0 10px 0;
}
.hbox {
  overflow-x: auto;
  white-space: nowrap;
}
.vbox {
  overflow-y: auto;
}
.intrinsic-height-box {
  min-height: -webkit-min-content;
  overflow: auto;
  flex-basis: 0;
}
.rect {
  min-height: 100px;
  min-width: 100px;
  background-color: green;
  display: inline-block;
}
</style>
<div class="vflex">
  <div class="hbox">
    <div class="rect"></div>
    <div class="rect"></div>
  </div>
</div>

<div class="vflex">
  <div class="intrinsic-height-box">
    <div class="rect" style="min-width: 300px"></div>
  </div>
</div>

<div class="hflex">
  <div class="vbox">
    <div class="rect" style="min-height: 300px;"></div>
  </div>
</div>

<div style="display: flex; width: 100px; height: 100px;">
  <div id="inner" style="flex: none; height: 100px; overflow: auto;">
    <div style="width: 100px; height: 150px; background-color: green;"></div>
  </div>
</div>

<!-- This div is only for measuring scrollbar size -->
<div id="measure" style="height: 100px; width: 100px; display: inline-box; overflow: auto;">
  <div style="min-height: 300px;"></div>
</div>

<script>
  description("When a block inside a flexbox adds scrollbars due to overflow, the parent flexbox should re-flex based on the child size including scrollbars.");

  var hbox = document.querySelector('.hbox');
  shouldBe("hbox.clientHeight", "hbox.scrollHeight");

  var intrinsicHeightBox = document.querySelector('.intrinsic-height-box');
  shouldBe("intrinsicHeightBox.clientHeight", "intrinsicHeightBox.scrollHeight");

  var measure = document.getElementById('measure');
  var scrollbarSize = measure.offsetWidth - measure.clientWidth;
  shouldNotBe("scrollbarSize", "0");

  var inner = document.getElementById('inner');
  shouldBe("inner.clientWidth", "100");
  shouldBe("inner.offsetWidth", "100 + scrollbarSize");

  var vbox = document.querySelector('.vbox');
  shouldBe("vbox.clientWidth", "100 - scrollbarSize");
</script>
