third_party/WebKit/LayoutTests/css3/flexbox/overflow-auto-resizes-correctly-expected.txt

When a block inside a flexbox adds scrollbars due to overflow, the parent flexbox should re-flex based on the child size including scrollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hbox.clientHeight is hbox.scrollHeight
PASS intrinsicHeightBox.clientHeight is intrinsicHeightBox.scrollHeight
PASS scrollbarSize is not 0
PASS inner.clientWidth is 100
PASS inner.offsetWidth is 100 + scrollbarSize
PASS vbox.clientWidth is 100 - scrollbarSize
PASS successfullyParsed is true

TEST COMPLETE
 
